refactor: dto 패키지 위치를 서비스 패키지 하위로 이동
