+++ src/main/java/roomescape/service/dto/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.service.dto;
 
 public record ReservationRequest(
         String name,
