+++ src/main/java/roomescape/service/dto/ReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.service.dto;
 
 public record ReservationTimeRequest(String startAt) {
 }
